move load command definition to pkg commands handler (closes #891)
